Modualizing the project
