+++ utils/renderToDom.js
@@ -1,0 +1,6 @@
+const renderToDOM = (divId, content) => {
+  const selectedDiv = document.querySelector(divId);
+  selectedDiv.innerHTML = content;
+};
+
+export default renderToDOM;
