+++ utils/studentArea.js
@@ -1,0 +1,10 @@
+import renderToDOM from './renderToDom';
+
+const studentAreas = () => {
+  const domString = `<div id="students">No Students</div>
+  <div id="voldy">No Death Eaters</div>`;
+
+  renderToDOM('#student-container', domString);
+};
+
+export default studentAreas;
